Crate hdf5_rs

Modules

prelude
The HDF5 prelude module.

Macros

h5call!

h5call_s!

h5lock!
Run an unsafe expression in a closure synchronized by a global reentrant mutex.

h5lock_s!
Run a safe expression in a closure synchronized by a global reentrant mutex.

h5try!

h5try_s!

Structs

Dataset
Represents the HDF5 dataset object.

Dataspace
Represents the HDF5 dataspace object.

File
Represents the HDF5 file object.

Filters
HDF5 filters and compression options.

Group
Represents the HDF5 group object.

Enums

Datatype
Represents the HDF5 datatype object.

Error
The error type for HDF5-related functions.

Traits

Container
A trait for HDF5 objects that can contain other objects (file, group).

Dimension
A trait for the shape and index types.

Location
A trait for HDF5 objects that can have a named location (file, group, dataset).

Object
A trait for all HDF5 objects that can be referenced through an identifier.

ToDatatype
A trait for native types that are convertible to HDF5 datatypes.

Functions

hdf5_version
Returns the version of the HDF5 library that the crate was compiled against.

Type Definitions

Ix
A scalar integer type used by Dimension trait for indexing.

Result
A type for results generated by HDF5-related functions where the Err type is set to hdf5::Error.