Module hdf5_rs::prelude

The HDF5 prelude module.

The purpose of this module is to provide reexports of many core hdf5 traits so that they can be then glob-imported all at once:

use hdf5_rs::prelude::*;

This module provides reexports of such traits as Object, Location and Container and does not expose any structures or functions.

Traits

AnyDatatype
A trait for all HDF5 datatypes.

AtomicDatatype
A trait for atomic scalar datatypes.

Container
A trait for HDF5 objects that can contain other objects (file, group).

Dimension
A trait for the shape and index types.

Location
A trait for HDF5 objects that can have a named location (file, group, dataset).

Object
A trait for all HDF5 objects that can be referenced through an identifier.

ToDatatype
A trait for native types that are convertible to HDF5 datatypes.