Trait hdf5_rs::prelude::Container

pub trait Container: Location {
    fn len(&self) -> u64 { ... }
    fn is_empty(&self) -> bool { ... }
    fn create_group(&self, name: &str) -> Result<Group> { ... }
    fn group(&self, name: &str) -> Result<Group> { ... }
    fn link_soft(&self, src: &str, dst: &str) -> Result<()> { ... }
    fn link_hard(&self, src: &str, dst: &str) -> Result<()> { ... }
    fn relink(&self, name: &str, path: &str) -> Result<()> { ... }
    fn unlink(&self, name: &str) -> Result<()> { ... }
    fn new_dataset<T: ToDatatype>(&self) -> DatasetBuilder<T> { ... }
    fn dataset(&self, name: &str) -> Result<Dataset> { ... }
}

A trait for HDF5 objects that can contain other objects (file, group).

Provided Methods

fn len(&self) -> u64

Returns the number of objects in the container (or 0 if the container is invalid).

fn is_empty(&self) -> bool

Returns true if the container has no linked objects (or if the container is invalid).

fn create_group(&self, name: &str) -> Result<Group>

Create a new group in a file or group.

fn group(&self, name: &str) -> Result<Group>

Opens an existing group in a file or group.

fn link_soft(&self, src: &str, dst: &str) -> Result<()>

Creates a soft link. Note: src and dst are relative to the current object.

fn link_hard(&self, src: &str, dst: &str) -> Result<()>

Creates a hard link. Note: src and dst are relative to the current object.

fn relink(&self, name: &str, path: &str) -> Result<()>

Relinks an object. Note: name and path are relative to the current object.

fn unlink(&self, name: &str) -> Result<()>

Removes a link to an object from this file or group.

fn new_dataset<T: ToDatatype>(&self) -> DatasetBuilder<T>

Instantiates a new dataset builder.

fn dataset(&self, name: &str) -> Result<Dataset>

Opens an existing dataset in the file or group.