Trait hdf5_rs::prelude::Location

pub trait Location: Object {
    fn name(&self) -> String { ... }
    fn filename(&self) -> String { ... }
    fn file(&self) -> Result<File> { ... }
    fn comment(&self) -> Option<String> { ... }
    fn set_comment(&self, comment: &str) -> Result<()> { ... }
    fn clear_comment(&self) -> Result<()> { ... }
}

A trait for HDF5 objects that can have a named location (file, group, dataset).

Provided Methods

fn name(&self) -> String

Returns the name of the object within the file, or empty string if the object doesn't have a name (e.g., an anonymous dataset).

fn filename(&self) -> String

Returns the name of the file containing the named object (or the file itself).

fn file(&self) -> Result<File>

Returns a handle to the file containing the named object (or the file itself).

fn comment(&self) -> Option<String>

Returns the commment attached to the named object, if any.

fn set_comment(&self, comment: &str) -> Result<()>

Set or the commment attached to the named object.

fn clear_comment(&self) -> Result<()>

Clear the commment attached to the named object.