Struct hdf5_rs::Dataset

pub struct Dataset {
    // some fields omitted
}

Represents the HDF5 dataset object.

Methods

impl Dataset

fn shape(&self) -> Vec<Ix>

Returns the shape of the dataset.

fn ndim(&self) -> usize

Returns the number of dimensions in the dataset.

fn size(&self) -> usize

Returns the total number of elements in the dataset.

fn is_scalar(&self) -> bool

Returns whether this dataset is a scalar.

fn is_resizable(&self) -> bool

Returns whether this dataset is resizable along some axis.

fn is_chunked(&self) -> bool

Returns whether this dataset has a chunked layout.

fn is_type<T: ToDatatype>(&self) -> bool

Returns whether this dataset's type is equivalent to the given type.

fn chunks(&self) -> Option<Vec<Ix>>

Returns the chunk shape if the dataset is chunked.

fn filters(&self) -> Filters

Returns the filters used to create the dataset.

fn tracks_times(&self) -> bool

Returns true if object modification time is tracked by the dataset.

fn storage_size(&self) -> u64

Returns the amount of file space required for the dataset. Note that this only accounts for the space which has actually been allocated (it can be equal to zero).

fn offset(&self) -> Option<u64>

Returns the absolute byte offset of the dataset in the file if such offset is defined (which is not the case for datasets that are chunked, compact or not allocated yet).

fn fill_value<T: ToDatatype>(&self) -> Result<Option<T>>

Returns default fill value for the dataset if such value is set. Note that conversion to the requested type is done by HDF5 which may result in loss of precision for floating-point values if the datatype differs from the datatype of of the dataset.

fn datatype(&self) -> Result<Datatype>

Returns a new Datatype object associated with this dataset.