Struct hdf5_rs::Filters

pub struct Filters {
    // some fields omitted
}

HDF5 filters and compression options.

Methods

impl Filters

fn new() -> Filters

fn gzip(&mut self, level: u8) -> &mut Filters

Enable gzip compression with a specified level (0-9).

fn no_gzip(&mut self) -> &mut Filters

Disable gzip compression.

fn get_gzip(&self) -> Option<u8>

Get the current settings for gzip filter.

fn szip(&mut self, nn: bool, level: u8) -> &mut Filters

Enable szip compression with a specified method (EC, NN) and level (0-32).

If nn if set to true (default), the nearest neighbor method is used, otherwise the method is set to entropy coding.

fn no_szip(&mut self) -> &mut Filters

Disable szip compression.

fn get_szip(&self) -> Option<(bool, u8)>

Get the current settings for szip filter.

Returns a tuple (nn, level), where nn indicates whether the nearest neighbor method is used and level is the associated compression level.

fn shuffle(&mut self, shuffle: bool) -> &mut Filters

Enable or disable shuffle filter.

fn get_shuffle(&self) -> bool

Get the current settings for shuffle filter.

fn fletcher32(&mut self, fletcher32: bool) -> &mut Filters

Enable or disable fletcher32 filter.

fn get_fletcher32(&self) -> bool

Get the current settings for fletcher32 filter.

fn scale_offset(&mut self, scale_offset: u32) -> &mut Filters

Enable scale-offset filter with a specified factor (0 means automatic).

fn no_scale_offset(&mut self) -> &mut Filters

Disable scale_offset compression.

fn get_scale_offset(&self) -> Option<u32>

Get the current settings for scale_offset filter.

fn gzip_default(&mut self) -> &mut Filters

Enable gzip filter with default settings (compression level 4).

fn szip_default(&mut self) -> &mut Filters

Enable szip filter with default settings (NN method, compression level 8).

fn has_filters(&self) -> bool

Returns true if any filters are enabled and thus chunkins is required.

fn validate(&self) -> Result<()>

Verify whether the filters configuration is valid.